Add per-file multithreading
When the program is given more than 1 filepath it handles these serially
1 at a time, this is leaving available disk, RAM and CPU resources
underutilized.

Make `handle_files()` spawn a thread for each file; this allows us to do
reads, writes, and text-processing (from different files) in parallel.
Which should translate to increased performance when handling multiple
files.

An alternative threading method that would cut each file into lines and
then distribute those among threads was considered: it was decided to be
too complicated, while it's still up for debate whether this would be
more performant or not; the chosen method is simple to implement, and a
guaranteed improvement.