Use `AsRef<Path>` and `Into<PathBuf>` instead of `OsString` and `OsStr`
The code uses OsString, this works but it could make it hard to work
with if the passed arguments' types change.

Use `AsRef<Path>` and `Into<PathBuf>` instead of `OsString` and `OsStr`;
this allows us to call these functions with a wide array of argument
types (without losing performance), which should ease future development
in case any of the argument types change (useful for writing tests).